% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_local_sum.R
\name{align_local_sum}
\alias{align_local_sum}
\title{summarize align local output}
\usage{
align_local_sum(al_var)
}
\arguments{
\item{al_var}{list produced by align local}
}
\value{
list
\itemize{
\item df dataframe of words in a an d
\item sum integer number of words the same
\item sentence string the sentence
}
}
\description{
summarize align local output
}
\examples{

}
