% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_1_col_pdf.R
\name{clean_1_col_pdf}
\alias{clean_1_col_pdf}
\title{clean load 1 column pdf}
\usage{
clean_1_col_pdf(file)
}
\arguments{
\item{file}{a file path}
}
\value{
character string of text
}
\description{
clean load 1 column pdf
}
\examples{

}
