% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_2_col_pdf.R
\name{clean_2_col_pdf}
\alias{clean_2_col_pdf}
\title{load and clean 2 column pdf}
\usage{
clean_2_col_pdf(file)
}
\arguments{
\item{file}{a file path}
}
\value{
character string of cleaned text
}
\description{
Note this function requires the tabulizer package, which also requires a java installation. It is a useful function for reading in two-column pdfs often found in the scientific literature.
}
\examples{

}
