% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_plain_txt.R
\name{clean_plain_txt}
\alias{clean_plain_txt}
\title{clean and load a plain text file}
\usage{
clean_plain_txt(file)
}
\arguments{
\item{file}{a path to file}
}
\value{
character string of text
}
\description{
clean and load a plain text file
}
\examples{

}
