% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorize_uppercase.R
\name{colorize_uppercase}
\alias{colorize_uppercase}
\title{Turn uppercase words red}
\usage{
colorize_uppercase(text)
}
\arguments{
\item{text}{string of text}
}
\value{
text string of text with css span color set to red for uppercase words
}
\description{
Turn uppercase words red
}
\examples{

}
