% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_doc_compare.R
\name{multi_doc_compare}
\alias{multi_doc_compare}
\title{Multiple document comparison for textual overlap}
\usage{
multi_doc_compare(texts, n_grams, sd_criterion)
}
\arguments{
\item{texts}{character vector of texts, each text is a string in the vector}

\item{n_grams}{integer to specify ngram units}

\item{sd_criterion}{numeric set a standard deviation criterion for returning documents that are unsually similar, 2-3 is pretty good}
}
\value{
list
\itemize{
\item dtm matrix document term matrix for all texts
\item histogram a histogram of the cosine similarity values between every text
\item similarities matrix cosine similarities between every text
\item mean_similarity numeric the mean similarity between all texts
\item sd_similarity numeric the standard deviation of the similarities
\item check_these dataframe document pairs that were above the criterion, might want to check these ones))
}
}
\description{
Multiple document comparison for textual overlap
}
\examples{

}
