% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngram_proportion_same.R
\name{ngram_proportion_same}
\alias{ngram_proportion_same}
\title{Proportion n gram overlap}
\usage{
ngram_proportion_same(a, b, n_grams)
}
\arguments{
\item{a}{character string of text}

\item{b}{character string of text}

\item{n_grams}{integer number of ngrams}
}
\value{
numeric proportion of ngram overlap
}
\description{
Proportion n gram overlap
}
\examples{
p1 <- "hello my name is someone someone"
p2 <- "my name is someone else"
ngram_proportion_same(p1,p2,2)
}
