% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngrams_analysis.R
\name{ngrams_analysis}
\alias{ngrams_analysis}
\title{Proportion ngram overlap for a range of ngrams}
\usage{
ngrams_analysis(a, b, range, meta = "none")
}
\arguments{
\item{a}{string of text}

\item{b}{string of text}

\item{range}{integer vector for ngram range (e.g., 3:5)}
}
\value{
list of lists, each list contains:
\itemize{
\item proportion numeric of overlapping ngrams
\item the_grams character vector of ngrams that overlap
\item a_print string the first document, with
}
}
\description{
Proportion ngram overlap for a range of ngrams
}
\examples{

}
