% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngrams_report.R
\name{ngrams_report}
\alias{ngrams_report}
\title{Produce ngram report in rmd for html}
\usage{
ngrams_report(out, print_n, highlight_n, color = "red")
}
\arguments{
\item{out}{list from ngrams_analysis}

\item{print_n}{integer number of ngram examples to print for each ngram type}

\item{highlight_n}{string "ngram4" the name of the ngram unit, for printing entire text for comparison, ngrams that are the same between texts are printed in red and uppercase.}
}
\value{
cat text which is printed below the knitr chunk
}
\description{
Produce ngram report in rmd for html
}
\examples{

}
