% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorize_uppercase.R
\name{colorize_uppercase}
\alias{colorize_uppercase}
\title{highlight uppercase words}
\usage{
colorize_uppercase(text, color = "red")
}
\arguments{
\item{text}{string of text}

\item{color}{a color code for html css}
}
\value{
text string of text with css span color set to red for uppercase words
}
\description{
This function detects uppercase words, then wraps them with an HTML span. These spans are given the id = ngram, and a color specified by color. The ngram id could be used in css to further customize the style of the highlighted text.
}
\examples{

}
